requirements-pipeline.txt
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
requests==2.31.0
redis==5.0.1
dvc==3.45.0
fairlearn==0.10.0
google-cloud-firestore==2.14.0
python-json-logger==2.0.7
apache-airflow==2.7.2
apache-airflow-providers-google==10.10.0
pytest==7.4.3
